feat(tests): add integration tests for GenkitProvider and mock server
Introduces comprehensive tests for the GenkitProvider using a mock AI backend.
This ensures no live API calls are made during testing while validating the
generation logic for different providers (Anthropic, OpenAI, Google).
